onboarding refined
