+++ clients/apple/iOS/Scanner.swift
@@ -1,0 +1,40 @@
+import SwiftUI
+
+struct Scanner: View {
+    
+    @EnvironmentObject var onboardingState: OnboardingService
+    
+    @State var isScanning: Bool = false
+    
+    var body: some View {
+        Button(action: {
+            self.isScanning = true
+        }) {
+            Image(systemName: "qrcode.viewfinder")
+        }
+        .frame(width: 40, height: 40)
+        .disabled(self.onboardingState.working)
+        .sheet(isPresented: $isScanning, content: {
+            CodeScannerView(codeTypes: [.qr], simulatedData: "This is simulated data", completion: handleScan)
+        })
+    }
+    
+    func handleScan(result: Result<String, CodeScannerView.ScanError>) {
+        self.isScanning = false
+        switch result {
+        case .success(let key):
+            self.onboardingState.accountString = key
+            self.onboardingState.handleImport()
+        case .failure(let err):
+            print(err) // TODO: Convert this to an ApplicationError
+        }
+    }
+}
+
+struct ImportView_Previews: PreviewProvider {
+    
+    static var previews: some View {
+        OnboardingView(selectedTab: .Import)
+            .mockDI()
+    }
+}
